+++ src/main/webapp/resource/js/back-table-tag.js
@@ -219,8 +219,13 @@ function deleteTags() {
 
     //$("#testJson").empty();
     //$("#testJson").append(JSON.stringify(ids));
-
-    deleteByIds(ids);
+    //debugger;
+    if (ids==""||ids==" ") {
+        document.getElementById("tipContent").innerText="请选择有效行";
+        $("#Tip").modal('show');
+    }else {
+        deleteByIds(ids);
+    }
 }
 
 function deleteByIds(ids) {
